+++ test/index.ts
@@ -1,53 +1,18 @@
-import WS from 'jest-websocket-mock';
 import { Agent } from '../dist/node';
-import path from 'path';
-import fs from 'fs';
-import { cacheValues, fillValues } from './utils/sip';
 
 export * from '../dist/node';
 
-export const getAgent = () => new Agent({
-  endpoint: 'ws://localhost:1234',
-  domain: 'dec112.at',
-  user: 'user',
-  password: 'password',
-  displayName: 'Alice',
-  // debug: true,
-});
-
-let _server: WS;
-
-const initialize = () => {
-  _server = new WS("ws://localhost:1234");
+export const getAgents = () => {
+  // this looks weird, but as it is a backport from ng112-js@2.0.0
+  // we just keep the structure of having more agents
+  // even if in version 1.0.0 there will always be only one agent.
+  return [
+    new Agent({
+      endpoint: 'ws://127.0.0.1:8088',
+      domain: 'service.dec112.home',
+      user: 'user',
+      password: '',
+      displayName: 'Alice Smith',
+    })
+  ];
 }
-
-const send = (filename: string) => {
-  const pathParts: string[] = [
-    __dirname,
-    'res',
-    ...filename.split('/'),
-  ]
-
-  let msg = fs.readFileSync(`${path.join(...pathParts)}.txt`, { encoding: 'utf-8' });
-  msg = fillValues(msg);
-
-  _server.send(msg);
-}
-
-const clean = () => WS.clean();
-
-const nextMessage = async () => {
-  let msg = (await _server.nextMessage) as string;
-  cacheValues(msg);
-  return msg;
-}
-
-export const server = {
-  send,
-  clean,
-  initialize,
-  is: () => _server,
-  expect: {
-    message: nextMessage,
-  },
-};