+++ test/integration/error.test.ts
@@ -1,0 +1,108 @@
+import { getAgents } from '..';
+import { Agent, Origin, ConversationState, StateObject, MessageFailedEvent } from '../../dist/node';
+import { initializeTests } from './utils';
+
+initializeTests();
+
+describe('ng112-js errors', () => {
+  it.each<Agent>(getAgents())('handles 404 Not Found errors correctly', async (agent: Agent) => {
+    const target = 'sip:non-existent@service.dec112.home';
+    await agent.initialize();
+
+    const conversation = agent.createConversation(target);
+    const positiveMock = jest.fn();
+
+    try {
+      await conversation.start().promise;
+      positiveMock();
+    } catch (_ex) {
+      const ex = _ex as MessageFailedEvent;
+
+      expect(ex.originator).toEqual(Origin.REMOTE);
+      expect(ex.response.status_code).toBe(404);
+    }
+
+    expect(positiveMock).not.toHaveBeenCalled();
+
+    await agent.dispose();
+  });
+
+  it.each<Agent>(getAgents())('handles errors with invalid domains correctly', async (agent: Agent) => {
+    const target = 'sip:default@domain.is.invalid';
+    await agent.initialize();
+
+    const conversation = agent.createConversation(target);
+    const positiveMock = jest.fn();
+
+    try {
+      await conversation.start().promise;
+      positiveMock();
+    } catch (_ex) {
+      const ex = _ex as MessageFailedEvent;
+
+      expect(ex.response.status_code).toBeGreaterThanOrEqual(400);
+      expect(ex.originator).toEqual(Origin.REMOTE);
+    }
+
+    expect(positiveMock).not.toHaveBeenCalled();
+
+    await agent.dispose();
+  });
+
+  it.each<Agent>(getAgents())('resumes a conversation that has run into an error', async (agent: Agent) => {
+    const target = 'sip:default@service.dec112.home';
+    await agent.initialize();
+
+    const conversation = agent.createConversation(target);
+
+    await conversation.start().promise;
+
+    // this is a private property, therefore typescript complains
+    // @ts-expect-error
+    conversation._targetUri = 'sip:non-existent@service.dec112.home';
+    const positiveMock = jest.fn();
+
+    try {
+      await conversation.sendMessage({
+        text: 'some text',
+      }).promise;
+      positiveMock();
+    } catch (_ex) {
+      const ex = _ex as MessageFailedEvent;
+      
+      expect(ex.originator).toEqual(Origin.REMOTE);
+      expect(ex.response.status_code).toBe(404);
+    }
+
+    expect(positiveMock).not.toHaveBeenCalled();
+
+    const errorState: StateObject = {
+      origin: Origin.REMOTE,
+      value: ConversationState.ERROR,
+    };
+    expect(conversation.state).toEqual(errorState);
+
+    // this is a private property, therefore typescript complains
+    // @ts-expect-error
+    conversation._targetUri = target;
+
+    // now we should be able to resume the conversation again
+    await conversation.sendMessage({
+      text: 'some text',
+    }).promise;
+
+    const okState: StateObject = {
+      origin: Origin.LOCAL,
+      value: ConversationState.STARTED,
+    };
+    expect(conversation.state).toEqual(okState);
+
+    await agent.dispose();
+
+    const stopState: StateObject = {
+      origin: Origin.LOCAL,
+      value: ConversationState.STOPPED,
+    };
+    expect(conversation.state).toEqual(stopState);
+  });
+});